### 2024-01-02 b66e004
move directional evaluation into stronge wolfe
