### 2024-01-07 f8d8786
remove unneeded comparison
