### 2024-01-07 c6f22cf
fix w decay in line search
