+++ src/lbfgs/strong_wolfe.rs
@@ -1,7 +1,6 @@
+use crate::Model;
 use candle_core::Result as CResult;
 use candle_core::{Tensor, Var};
-
-use crate::Model;
 
 use super::{add_grad, flat_grads, set_vs, Lbfgs};
 
@@ -91,7 +90,7 @@ impl<M: Model> Lbfgs<M> {
             .to_scalar::<f64>()?;
 
         // evaluate objective and gradient using initial step
-        let (f_new, g_new) = self.directional_evaluate(step_size, direction)?;
+        let (f_new, g_new, mut l2_new) = self.directional_evaluate(step_size, direction)?;
         let g_new = Var::from_tensor(&g_new)?;
         let f_new = Var::from_tensor(&f_new)?;
         let mut ls_func_evals = 1;
@@ -106,22 +105,28 @@ impl<M: Model> Lbfgs<M> {
         // bracket an interval containing a point satisfying the Wolfe criteria
         let g_prev = Var::from_tensor(grad)?;
         let f_prev = Var::from_tensor(loss)?;
+        let l2_init = self.l2_reg()?;
+        let mut l2_prev = l2_init;
         let (mut t_prev, mut gtd_prev) = (0., directional_grad);
         let mut done = false;
         let mut ls_iter = 0;
 
         let mut bracket_gtd;
+        let mut bracket_l2;
         let bracket_f;
         let (mut bracket, bracket_g) = loop {
             // check conditions
             if f_new
                 .to_dtype(candle_core::DType::F64)?
                 .to_scalar::<f64>()?
+                + l2_new
                 >= f_prev
                     .to_dtype(candle_core::DType::F64)?
                     .to_scalar::<f64>()?
+                    + l2_prev
             {
                 bracket_gtd = [gtd_prev, gtd_new];
+                bracket_l2 = [l2_prev, l2_new];
                 bracket_f = [f_prev, Var::from_tensor(f_new.as_tensor())?];
                 break (
                     [t_prev, step_size],
@@ -132,6 +137,7 @@ impl<M: Model> Lbfgs<M> {
             if gtd_new.abs() <= -c2 * directional_grad {
                 done = true;
                 bracket_gtd = [gtd_prev, gtd_new];
+                bracket_l2 = [l2_prev, l2_new];
                 bracket_f = [
                     Var::from_tensor(f_new.as_tensor())?,
                     Var::from_tensor(f_new.as_tensor())?,
@@ -147,6 +153,7 @@ impl<M: Model> Lbfgs<M> {
 
             if gtd_new >= 0. {
                 bracket_gtd = [gtd_prev, gtd_new];
+                bracket_l2 = [l2_prev, l2_new];
                 bracket_f = [f_prev, Var::from_tensor(f_new.as_tensor())?];
                 break (
                     [t_prev, step_size],
@@ -162,12 +169,14 @@ impl<M: Model> Lbfgs<M> {
                 t_prev,
                 f_prev
                     .to_dtype(candle_core::DType::F64)?
-                    .to_scalar::<f64>()?,
+                    .to_scalar::<f64>()?
+                    + l2_prev,
                 gtd_prev,
                 step_size,
                 f_new
                     .to_dtype(candle_core::DType::F64)?
-                    .to_scalar::<f64>()?,
+                    .to_scalar::<f64>()?
+                    + l2_new,
                 gtd_new,
                 Some((min_step, max_step)),
             );
@@ -176,12 +185,15 @@ impl<M: Model> Lbfgs<M> {
             t_prev = tmp;
             f_prev.set(f_new.as_tensor())?;
             g_prev.set(g_new.as_tensor())?;
+            l2_prev = l2_new;
             gtd_prev = gtd_new;
             // assign to temp vars:
-            let (next_f, next_g) = self.directional_evaluate(step_size, direction)?;
[…]
