### 2024-01-07 b874db1
add test for l2 reg
