+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "candle-optimisers"
-version = "0.3.1"
+version = "0.3.2"
 edition = "2021"
 readme = "README.md"
 license = "MIT"
