### 2024-02-09 #6167 @KGrewal1
Would it be possible to bump the version to v0.3.2 : there would be no changes to the paper's content, with the change being a correctness fix in the LBFGS linesearch when used with weight decay?
