### 2024-02-09 #6167 @olexandr-konovalov
@KGrewal1 No need to do it now, I think, but the version reported in the paper is checked and updated if necessary upon its acceptance (see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance).
