+++ Cargo.toml
@@ -19,7 +19,6 @@ exclude = [
 
 candle-core = { git = "https://github.com/huggingface/candle.git", version = "0.3.0"}
 candle-nn = { git = "https://github.com/huggingface/candle.git", version = "0.3.0"}
-intel-mkl-src = { version = "0.8.1", features = ["mkl-static-lp64-iomp"], optional = true }
 log = "0.4.20"
 
 
@@ -37,7 +36,6 @@ harness = false
 [features]
 default = []
 cuda = ["candle-core/cuda", "candle-nn/cuda"]
-mkl = ["dep:intel-mkl-src", "candle-core/mkl"]
 
 [profile.bench]
 lto = true # maximal LTO optimisaiton
