### 2024-02-28 521be46
update candle crates to v0.4
