+++ Changelog.md
@@ -3,6 +3,7 @@
 ## v0.4.0 (2024-02-28)
 
 * Bump candle requirtement to 0.4.0: this is considered a breaking change due to the reliance of this library on candle-core and candle-nn
+* Explicit reliance on the candle crates hosted on crates.io : as cargo does not support git dependecies in published crates, this library now points only to the crates.io releases (previously cargo would default to the crates.io instead of git repo anyway: if the git repo is specifically desired this can be obtained by patching the `Cargo.toml` to point at the candle repo)
 * Remove intel-mkl feature: features in this library are mainly used for running examples: any code that uses this library should instead use the features directly from the candle crates
 
 ## v0.3.2 (2024-01-07)
