+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -8,13 +8,10 @@ assignees: ''
 ---
 
 **Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+Is the feature an additional optimisation algorithm?  Is it a change to the documentation?
 
 **Describe the solution you'd like**
-A clear and concise description of what you want to happen.
-
-**Describe alternatives you've considered**
-A clear and concise description of any alternative solutions or features you've considered.
+A clear and concise description of what you want to happen: for additional optimisation algorithms it would be ideal if possible to see if there are other implementations or a reference (though am very happy to accept novel optimisers).
 
 **Additional context**
-Add any other context or screenshots about the feature request here.
+Add any other context  about the feature request here.
