### 2024-05-03 ad16408
change f to scalar in line search
