+++ src/lbfgs/strong_wolfe.rs
@@ -92,7 +92,9 @@ impl<M: Model> Lbfgs<M> {
         // evaluate objective and gradient using initial step
         let (f_new, g_new, mut l2_new) = self.directional_evaluate(step_size, direction)?;
         let g_new = Var::from_tensor(&g_new)?;
-        let f_new = Var::from_tensor(&f_new)?;
+        let mut f_new = f_new
+            .to_dtype(candle_core::DType::F64)?
+            .to_scalar::<f64>()?;
         let mut ls_func_evals = 1;
         let mut gtd_new = g_new
             .unsqueeze(0)?
@@ -104,7 +106,10 @@ impl<M: Model> Lbfgs<M> {
 
         // bracket an interval containing a point satisfying the Wolfe criteria
         let g_prev = Var::from_tensor(grad)?;
-        let f_prev = Var::from_tensor(loss)?;
+        let dtype = loss.dtype();
+        let shape = loss.shape();
+        let dev = loss.device();
+        let mut f_prev = loss.to_dtype(candle_core::DType::F64)?.to_scalar::<f64>()?;
         let l2_init = self.l2_reg()?;
         let mut l2_prev = l2_init;
         let (mut t_prev, mut gtd_prev) = (0., directional_grad);
@@ -113,21 +118,13 @@ impl<M: Model> Lbfgs<M> {
 
         let mut bracket_gtd;
         let mut bracket_l2;
-        let bracket_f;
+        let mut bracket_f;
         let (mut bracket, bracket_g) = loop {
             // check conditions
-            if f_new
-                .to_dtype(candle_core::DType::F64)?
-                .to_scalar::<f64>()?
-                + l2_new
-                >= f_prev
-                    .to_dtype(candle_core::DType::F64)?
-                    .to_scalar::<f64>()?
-                    + l2_prev
-            {
+            if f_new + l2_new >= f_prev + l2_prev {
                 bracket_gtd = [gtd_prev, gtd_new];
                 bracket_l2 = [l2_prev, l2_new];
-                bracket_f = [f_prev, Var::from_tensor(f_new.as_tensor())?];
+                bracket_f = [f_prev, f_new];
                 break (
                     [t_prev, step_size],
                     [g_prev, Var::from_tensor(g_new.as_tensor())?],
@@ -138,10 +135,7 @@ impl<M: Model> Lbfgs<M> {
                 done = true;
                 bracket_gtd = [gtd_prev, gtd_new];
                 bracket_l2 = [l2_prev, l2_new];
-                bracket_f = [
-                    Var::from_tensor(f_new.as_tensor())?,
-                    Var::from_tensor(f_new.as_tensor())?,
-                ];
+                bracket_f = [f_new, f_new];
                 break (
                     [step_size, step_size],
                     [
@@ -154,7 +148,7 @@ impl<M: Model> Lbfgs<M> {
             if gtd_new >= 0. {
                 bracket_gtd = [gtd_prev, gtd_new];
                 bracket_l2 = [l2_prev, l2_new];
-                bracket_f = [f_prev, Var::from_tensor(f_new.as_tensor())?];
+                bracket_f = [f_prev, f_new];
                 break (
                     [t_prev, step_size],
                     [g_prev, Var::from_tensor(g_new.as_tensor())?],
@@ -167,23 +161,17 @@ impl<M: Model> Lbfgs<M> {
             let tmp = step_size;
             step_size = cubic_interpolate(
                 t_prev,
-                f_prev
-                    .to_dtype(candle_core::DType::F64)?
-                    .to_scalar::<f64>()?
-                    + l2_prev,
+                f_prev + l2_prev,
                 gtd_prev,
                 step_size,
-                f_new
-                    .to_dtype(candle_core::DType::F64)?
-                    .to_scalar::<f64>()?
[…]
