+++ Cargo.toml
@@ -17,15 +17,15 @@ exclude = [
 
 [dependencies]
 
-candle-core =   "0.4.0"
-candle-nn = "0.4.0"
+candle-core =   "0.5.0"
+candle-nn = "0.5.0"
 log = "0.4.20"
 
 
 [dev-dependencies]
 anyhow = { version = "1", features = ["backtrace"] }
 assert_approx_eq = "1.1.0"
-candle-datasets = "0.4.0"
+candle-datasets = "0.5.0"
 clap = {version = "4.4.6", features = ["derive"] }
 criterion = { version = "0.5.1", features = ["html_reports"] }
 
