### 2024-05-04 f4c208f
clippy lint setup
