### 2024-05-22 #6167 @olexandr-konovalov
@relf and @kno10 would one or both of you be able to review this submission for JOSS, please?

If not, could you possibly recommend someone whom we may approach regarding this?
