### 2024-05-22 #6167 @kno10
I know little about optimizers for neural networks, as I do not use neural networks in my research (hence I do not even know whom to recommend instead - anyone from torch maybe?).
From a brief look, I would argue that the claim of "lower overhead" should be somehow demonstrated.

Also, I do not see the point of having these separately published from candle itself if they cannot be used without. Optimizers seems to be a must-have for me for any neural network library, so it seems an integral part of a larger package. For software architecture its nice to have them in a separate crate, but I don't see it as standalone thing.
