### 2024-05-22 #6167 @relf
Hi @olexandr-konovalov, sure I would be happy to help. I am not familiar with gradient optimization and neural networks but I can run examples, tests and take a look at the paper, the software quality and the documentation.
