### 2024-05-22 #6167 @olexandr-konovalov
Hi @jatkinson1000 and @anilbey - we are looking for a second reviewer of this paper, who is familiar its application domain (gradient optimization and neural networks). Would one of you be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?
