### 2024-05-23 #6167 @relf
@olexandr-konovalov, May I suggest the developer/maintainer of [argmin](https://github.com/argmin-rs/argmin) as a reviewer? `argmin` is, to the best of my knowledge, the first solution when one is seeking for optimizers in the Rust ecosystem.

As a first pass, the paper should cite existing libraries and highlight the contributions in comparison to those libraries (at least wrt current optimizers in `candle` and `argmin`).
