### 2024-05-23 #6167 @olexandr-konovalov
Hi @stefan-k, we would be very grateful if you'd be able to review this submission for JOSS - please see the comments above for the background. If not, could you possibly recommend someone else whom we may approach?
