### 2024-05-29 #6167 @jatkinson1000
Hi @olexandr-konovalov I was on leave last week.
I am potentially able to look at this. I'd note that my Rust is _extremely_ basic, though it sounds like you are after someone from the applications side for reviewer 2.

That said, I will be away at a conference for the next 2 weeks so there would be some delay before I could look at it, so if you find an alternative I am more than happy for them to take it.

On a quick glance I would echo others' comments about justifying performance claims, and demonstrating coupling/use in multiple applications to justify why this should not be directly integrated into `candle`.
