### 2024-05-29 #6167 @KGrewal1
Hi, read through some of the comments: namely re integration into `candle`, and that is due to the choice of the `candle` maintainers choosing to keep the optimisers outside of `candle` itself. As this is my first submission to JOSS I appreciate this may not be in the scope of JOSS as it seems it is too minimal
