### 2024-05-30 #6167 @jatkinson1000
Hi @KGrewal1 whilst I am not assigned to review, I would say that that is a valid rebuttal and should be made clear in the paper. Whilst contribution is generally encouraged over reinvention/large dependency trees, sometimes there are good reasons for alternatives.
Given this, and the fact that, well, you kind of need optimisers to run a deep learning model, I would personally feel that this _is_ appropriate for JOSS, though @olexandr-konovalov would be best to advise as editor.
