### 2024-06-04 #6167 @olexandr-konovalov
Hello everyone and thank you for the replies. We still haven't heard back from other invited reviewers, so may I suggest that if we will not hear until mid-June, I will start the review with @relf and @jatkinson1000 - I hope that together you will make a good coverage of all needed aspects, and as [editorial guide says](https://joss.readthedocs.io/en/latest/editing.html#adding-a-new-reviewer-once-the-review-has-started) we can add a new reviewer later too, if need be.

@KGrewal1, thanks for the explanations. I think that the paper fits the scope of the journal, and I agree with the suggestions expressed here so far. I hope those will be elaborated further and addressed when the review starts!
