### 2024-06-19 #6903 @editorialbot
[REVIEW]: Candle Optimisers: A Rust crate for optimisation algorithms
**Submitting author:** @KGrewal1 (<a href="http://orcid.org/0009-0001-7923-9975">Kirpal Grewal</a>)
**Repository:** https://github.com/KGrewal1/optimisers
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.1
