### 2024-06-21 KGrewal1/candle-optimisers#14 @relf
[JOSS Review - relf] Repo and README
As a first pass for https://github.com/openjournals/joss-reviews/issues/6903, my suggestions:

1. Repo name: Would it be too late to rename it `candle-optimizers`? I saw that the trait from `candle` is `Optimizer` with a `z`.
2. Installation: it could help to mention the main dependances `candle-core`, `candle-nn` maybe with the relevant versions along with `candle-optimizers` (something like a `Cargo.toml` snippet to include).
3. I run the example without problem, maybe you can highlight the presence of options (`--help`).
4. I did not succeed in using the instructions in `Notes`. I understand you want to mirror `pytorch.optim` capability (like `candle` mirrors `pytorch`), maybe you can state this as a goal for the crate.
5. It would be nice to position this crate with regards to `candle` optimization capabilities (`Òptimizer` trait and its current internal optimizers) and [argmin](https://github.com/argmin-rs/argmin) which is a rather popular optimizer framework in the Rust ecosystem (maybe some `argmin` optimizers could be adapted for working with `candle`?).
