### 2024-06-21 KGrewal1/optimisers#14 @KGrewal1
The change from s to z isn't really possible now (as it is published on crates.io and I don't believe the name change is possible there). Was going to change the repo name but unsure how that interacts with people using it as a dependency from GitHub directly (this should be a non issue I hope tho). For the rest, will try to fix these when I am able to
