### 2024-06-21 KGrewal1/optimisers#14 @relf
For crate.io, I would suggest to "reserve" the name by publishing at first minimal fake crate 'candle-optimizers' with just a README explaining the renaming transition. You would be able to deprecate the old crate in favour of the new one when it is ready.
