### 2024-06-21 KGrewal1/optimisers#14 @KGrewal1
I'm just not really sure about the value add from the change from British to American spelling (though agree should be prefaced candle-)
