### 2024-06-21 KGrewal1/optimisers#14 @relf
Well, it is for the consistency with the ecosystem as "optimizer" is used in candle, pytorch... And wherever I see it in scipy, scikit-learn, etc. But in the end it is your call, not blocking for me.
