### 2024-07-13 b474292
bump to new candle
