+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 readme = "README.md"
 license = "MIT"
 keywords = ["optimisers", "candle", "tensor", "machine-learning"]
-categories = ["science", "machine-learning", "optimisation"]
+categories = ["science"]
 description = "Optimisers for use with candle, the minimalist ML framework"
 repository = "https://github.com/KGrewal1/optimisers"
 exclude = ["*.ipynb"]
