### 2024-08-23 #6903 @relf
To the best of my ability, I've completed my review of this library. I've installed and run the software test suite and the example
successfully. The documentation is in place using standard Rust doc tool and contains algorithms pseudo-code and reference papers.

The paper itself should elaborate on the "far lower overhead for neural network training" claim which is not so obvious to me.
I would not call PyTorch or Tensorflow "full Python frameworks", as actually the training heavylifting is done using
compiled languages under the hood. It should also consolidate the position of the library with regard to the existing
Rust landscape at least by referencing current `candle` optimizers and established optimizers crate like `argmin`.
