### 2024-09-03 #6903 @jatkinson1000
Hi All,

* I would reiterate the comments from the [pre-review phase](https://github.com/openjournals/joss-reviews/issues/6167):
  * Performance claims should be backed up
  * Include an explanation as to why this is not a contribution directly to Candle
  * Citations of other relevant/alternative libraries should be included and discussed
* The paper itself is a statement of need and a list of implemented algorithms. I would expect more discussion of things such as:
  * How to interface this library with Candle
  * How the library is structured, and an overview of the API
  * Examples of use
  * Demonstration of performance claims
* I would like the documentation to be explicitly linked from the README (and the paper) - It took me a long time to find it by clicking the 'docs: passing' badge.
* The paper says that this provides a "complete" set of optimisers "for Rust", but the repository suggests they are specifically for use with candle, and there is a list of unimplemented methods in a TODO list. This needs clarifying and making consistent.
* A software comment on tests would be that the coverage of the library is high, but largely consists of integration tests for each method comparing to previous values or comparative values from PyTorch. It would be good to see some unit tests in there as well, and some wider parameter sweeps on the integration tests.

In summary, whilst the software repository looks good the paper needs more work and content.
I would also like to see more information in the README, though see @relf has started [an issue](https://github.com/KGrewal1/candle-optimisers/issues/14) for this.
