### 2024-10-04 #6903 @olexandr-konovalov
@relf @jatkinson1000 - many thanks for the review and the summary of suggestions!!! We also encourage to post more lengthy suggestions or requests as separate issues directly in the submission repository, linking them back to the review issue. That helps to keep them individually tracked, and is useful both to the author and to the reviewer.

@KGrewal1 it seems like more work is needed both on the software and the paper itself, so it's now over to you. Please interact with the reviewers and try to address their suggestions.
