### 2024-10-05 #6903 @KGrewal1
Hi very sorry for the slow replies - had an internship during which I wasn't allowed to easily do open source stuff. I agree with the performance claim criticism and will remove that (as seems to be at best conditionally true depending on architecture).  Will try to add in more information about the API (this will probably be into the rust documentation rather than the paper as that seems a more natural place for it to go).

Regarding other optimisation libraries, the comparisons were the code in the candle main repo and burn - argmin whilst useful for general minimisation problems doesn't seem well suited for this comparison as it lacks stochastic optimisers - the aim was to be in a similar place to pytorch for ML training in rust.

Re not including in the main candle repo, this was to reduce maintenance of the core so that could move at a different speed - see https://github.com/huggingface/candle/pull/1171#issuecomment-1777850810
