### 2024-10-05 47cce68
remove notes for being confusing
