### 2024-10-05 5f2d993
add Adam example to docs
