+++ .github/workflows/codecov.yml
@@ -18,8 +18,8 @@ jobs:
         uses: actions/checkout@v3
       - name: Generate code coverage
         uses: dtolnay/rust-toolchain@stable
-      - run: |
-          cargo tarpaulin --verbose --exclude-files benches/* --out xml
+      - run: cargo clean && cargo install cargo-tarpaulin
+      - run: cargo tarpaulin --verbose --exclude-files benches/* --out xml
       - name:                   Upload to codecov.io
         uses:                   codecov/codecov-action@v3
         with:
