### 2024-10-05 1748f75
fix binstall
