+++ .github/workflows/codecov.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install binstall
         run: curl -L --proto '=https' --tlsv1.2 -sSf https://raw.githubusercontent.com/cargo-bins/cargo-binstall/main/install-from-binstall-release.sh | bash
       - name: Install cargo-tarpaulin
-        run: cargo clean && cargo binstall cargo-tarpaulin
+        run: cargo clean && cargo binstall cargo-tarpaulin -y
       - name: Run tests with coverage
         run: cargo tarpaulin --verbose --exclude-files benches/* --out xml
       - name:                   Upload to codecov.io
