+++ Cargo.toml
@@ -15,15 +15,15 @@ exclude = ["*.ipynb"]
 
 [dependencies]
 
-candle-core = "0.7.2"
-candle-nn = "0.7.2"
+candle-core = "0.8.0"
+candle-nn = "0.8.0"
 log = "0.4.20"
 
 
 [dev-dependencies]
 anyhow = { version = "1", features = ["backtrace"] }
 assert_approx_eq = "1.1.0"
-candle-datasets = "0.7.2"
+candle-datasets = "0.8.0"
 clap = { version = "4.4.6", features = ["derive"] }
 criterion = { version = "0.5.1", features = ["html_reports"] }
 
