### 2024-12-09 #6903 @olexandr-konovalov
Hi @KGrewal1 - I have seen that you have made a new release, but in the meantime the paper had only one commit with a few changes. Just to check, is that correct to assume that you are still working on the the next version of the paper?
