### 2024-12-09 #6903 @KGrewal1
The new release was to ensure it was kept up to date with the candle library - from comments I'm not so sure how to progress with the paper as may just be it is too limited in scope for JOSS
