### 2025-02-13 #6903 @crvernon
:wave: @KGrewal1 - I am the associate editor-in-chief for this track and was reading through this thread a bit.  Could you clarify what you mean by "from comments I'm not so sure how to progress with the paper as may just be it is too limited in scope for JOSS" which you stated above?  Are you saying that you will not be able to continue with this submission due to the inability to respond to what the reviewers have posed as issues?  Thanks.
